Fix: ensure that setting socketaddress family works on macOS
On macOS sa_family is uint8.

+++ tests/quic/helpers/path.nim
@@ -2,7 +2,7 @@ import ngtcp2
 import nativesockets
 
 var zeroAddress: Sockaddr
-zeroAddress.sa_family = AF_INET.uint16
+zeroAddress.sa_family = AF_INET.uint8
 
 let zeroPath* = ngtcp2_path(
   local: ngtcp2_addr(`addr`: addr zeroAddress, addrlen: sizeof(SockAddr).uint),
